+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 ### Added
 - Sources: new 'All playlists' type source available for playlist sources, similar to Active playlist or Playing playlist but tracking all available playlists (Current or new ones). i.e. it serves as a playlists content viewer.
 - Sources: new setting 'Multi-branch playlist view' at Views submenu to split tree by playlist's name for playlist(s) sources. i.e. if 2 playlists are used as source, tracks are separated into 2 branches by playlist instead of mixed in the tree.
+- Drag n' drop: new drag n' drop action for playlist sources to easily move/add tracks from/to playlists. While tracking active/playing playlist, tracks can be added to it from any other panel; for 'All playlists' source, tracks can also be moved within the panel from one playlist to another while using 'Multi-branch playlist view' setting. Note tracks are always added to the bottom of the playlist in any case.
 ### Changed
 - UI: improved mouse cursor effects in some cases.
 - Sources: when selecting multiple playlist as sources, by pressing Shift, if a playlist is already included it will be removed instead. Additionally duplicated playlists (by name or GUID) are no longer allowed.